change mnist reader

+++ ass_4_nearest_neighbor/ass4_mnist/mnist.h
@@ -7,7 +7,7 @@
 
 using namespace std;
 using namespace caffe;
-typedef int Dtype;
+typedef double Dtype;
 
 class MNIST
 {
